use read_input in compare method.
This required a handful of changes.  The earlier versions used
IO.foreach in a few places; since that apparently balks if you throw it
a string, I had to replace it.

I think the changes are for the best.  E.g., I think
    count1 = file.lines.count
is preferable to
    count1 = File.foreach(file1).inject(0) {|c, line| c+1}
Though I could be wrong about this (I only made the change because I had
to!).